perf: Clean local manifests conditionally

+++ android/build.sh
@@ -7,7 +7,10 @@ fi
 echo "--- Syncing"
 
 cd /buildkite/src
-rm -rf .repo/local_manifests/*
+grep -q $DEVICE out/last_device
+if [ $? -gt 0 ]; then
+    rm -rf .repo/local_manifests/*
+fi
 repo init -u https://github.com/Project-1CE/manifest -b sugisawa --depth 1
 # repo forall -vc "git clean -fdx; git checkout -f"
 repo sync -c -j$(nproc --all) --force-sync --fail-fast --no-clone-bundle --no-tags || exit 1
